improvements to listeners in service
